tests/requirements.txt
# Test Suite Dependencies

python-dotenv>=1.0.0
openai>=1.0.0
colorama>=0.4.6
